+++ src/Libraries/Microsoft.Extensions.AmbientMetadata.Application/ApplicationMetadataConfigurationBuilderExtensions.cs
@@ -20,8 +20,8 @@ public static class ApplicationMetadataConfigurationBuilderExtensions
     /// </summary>
     /// <param name="builder">The configuration builder.</param>
     /// <param name="hostEnvironment">An instance of <see cref="IHostEnvironment" />.</param>
-    /// <param name="sectionName">Section name to save configuration into. Default set to "ambientmetadata:application".</param>
-    /// <returns>The value of <paramref name="builder"/>>.</returns>
+    /// <param name="sectionName">The section name to save configuration into. The default is "ambientmetadata:application".</param>
+    /// <returns>The value of <paramref name="builder"/>.</returns>
     /// <exception cref="ArgumentNullException"><paramref name="builder"/> or <paramref name="hostEnvironment"/> is <see langword="null"/>.</exception>
     /// <exception cref="ArgumentException"><paramref name="sectionName"/> is either <see langword="null"/>, empty, or whitespace.</exception>
     public static IConfigurationBuilder AddApplicationMetadata(this IConfigurationBuilder builder, IHostEnvironment hostEnvironment, string sectionName = DefaultSectionName)