+++ src/Libraries/Microsoft.Extensions.AI.AzureAIInference/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History
 
+## 9.10.0-preview.1.25513.3
+
+- Updated to accommodate the additions in `Microsoft.Extensions.AI.Abstractions`.
+
 ## 9.9.1-preview.1.25474.6
 
 - Updated to accommodate the additions in `Microsoft.Extensions.AI.Abstractions`.